requirements.txt
pandas>=2.0.0
pyarrow>=14.0.0
pyyaml>=6.0.0
numpy>=1.24.0
